% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{JAJ_demo}
\alias{JAJ_demo}
\title{Demo JAJ}
\usage{
JAJ_demo(num_items = 3L, take_training = TRUE,
  feedback = JAJ::JAJ_feedback_with_score(), admin_password = "demo",
  researcher_email = "longgold@gold.uc.ak", dict = JAJ::JAJ_dict,
  language = "EN", ...)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{take_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults to a graph-based feedback page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{kf@omniversum.de},
the email address of this package's developer.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"EN"}), German (\code{"DE"})
and Russion (\code{"RU"}).
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{JAJ}()}.}
}
\description{
This function launches a demo for the JAJ.
}
