% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{JAJ_feedback_with_score}
\alias{JAJ_feedback_with_score}
\title{JAJ feedback (with score)}
\usage{
JAJ_feedback_with_score(dict = JAJ::JAJ_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textuak  feedback at the end of the test.
}
\examples{
\dontrun{
demo_JAJ(feedback = JAJ_feedback_with_score())}

}
