% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{JAJ_standalone}
\alias{JAJ_standalone}
\title{Standalone JAJ}
\usage{
JAJ_standalone(
  title = NULL,
  num_items = 16L,
  with_feedback = FALSE,
  with_welcome = TRUE,
  take_training = TRUE,
  admin_password = "conifer",
  researcher_email = "longgold@gold.uc.ak",
  languages = c("en", "de", "de_f", "ru", "it", "es", "lv"),
  dict = JAJ::JAJ_dict,
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items for the test.}

\item{with_feedback}{(Boolean scalar) Defines whether feedback about test performance shall be given
at the end of the test.
Defaults to TRUE.}

\item{with_welcome}{(Logical scalar) Whether to include a welcome message. Defaults to TRUE.}

\item{take_training}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}), German (\code{"de"}), formal German (\code{"de_f"}),
Italian (\code{"it"}), Spanish \code{"es"}, Latvian \code{"lv"}, and Russian (\code{"ru"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation. Defaults to  JAJ_dict.}

\item{validate_id}{An external function for validating IDs, which takes a string ID as input and returns a BOOELAN.
Defaults to "auto", which validates purely alphanumeric IDs with no more than 100 characters.}

\item{...}{Further arguments to be passed to \code{\link{JAJ}()}.}
}
\description{
This function launches a standalone testing session for the MPT.
This can be used for data collection, either in the laboratory or online.
}
