% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAJ.R
\name{JAJ}
\alias{JAJ}
\title{JAJ}
\usage{
JAJ(num_items = 16L, take_training = TRUE, label = "JAJ",
  feedback = JAJ_feedback_with_score(), next_item.criterion = "bOpt",
  next_item.estimator = "BM", next_item.prior_dist = "norm",
  next_item.prior_par = c(0, 1), final_ability.estimator = "WL",
  dict = JAJ::JAJ_dict)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{take_training}{(Logical scalar) Whether to include the training phase.}

\item{label}{(Character scalar) Label to give the MPT results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults JAJ_feedback_with_score.}

\item{next_item.criterion}{(Character scalar)
Criterion for selecting successive items in the adaptive test.
See the \code{criterion} argument in \code{\link[catR]{nextItem}} for possible values.
\code{"bOpt"} corresponds to the setting used in the original JAJ paper.}

\item{next_item.estimator}{(Character scalar)
Ability estimation method used for selecting successive items in the adaptive test.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
\code{"BM"}, Bayes modal,
corresponds to the setting used in the original JAJ paper.
\code{"WL"}, weighted likelihood,
corresponds to the default setting used in versions <= 0.2.0 of this package.}

\item{next_item.prior_dist}{(Character scalar)
The type of prior distribution to use when calculating ability estimates
for item selection.
Ignored if \code{next_item.estimator} is not a Bayesian method.
Defaults to \code{"norm"} for a normal distribution.
See the \code{priorDist} argument in \code{\link[catR]{thetaEst}} for possible values.}

\item{next_item.prior_par}{(Numeric vector, length 2)
Parameters for the prior distribution;
see the \code{priorPar} argument in \code{\link[catR]{thetaEst}} for details.
Ignored if \code{next_item.estimator} is not a Bayesian method.
The dfeault is \code{c(0, 1)}.}

\item{final_ability.estimator}{Estimation method used for the final ability estimate.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
The default is \code{"WL"}, weighted likelihood,
which corresponds to the setting used in the original JAJ paper.
If a Bayesian method is chosen, its prior distribution will be defined
by the \code{next_item.prior_dist} and \code{next_item.prior_par} arguments.}

\item{dict}{The psychTestR dictionary used for internationalisation. Defaults to JAJ_dict}
}
\description{
This function defines a JAJ module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MPT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For demoing the JAJ, consider using \code{\link{JAJ_demo}()}.
For a standalone implementation of the MPT,
consider using \code{\link{JAJ_standalone}()}.
}
